backend/requirements.txt
fastapi[standard]
groq 
pydantic-settings
python-dotenv
tavily-python
sqlalchemy
asyncpg
openai
anthropic